+++ app/sprinkles/account/tests/Integration/AuthenticatorTest.php
@@ -12,8 +12,10 @@ namespace UserFrosting\Sprinkle\Account\Tests\Integration;
 use UserFrosting\Sprinkle\Account\Authenticate\Authenticator;
 use UserFrosting\Sprinkle\Account\Facades\Password;
 use UserFrosting\Sprinkle\Account\Tests\withTestUser;
+use UserFrosting\Sprinkle\Core\Database\Models\Session as SessionTable;
 use UserFrosting\Sprinkle\Core\Tests\TestDatabase;
 use UserFrosting\Sprinkle\Core\Tests\RefreshDatabase;
+use UserFrosting\Sprinkle\Core\Tests\withDatabaseSessionHandler;
 use UserFrosting\Tests\TestCase;
 
 /**
@@ -26,6 +28,7 @@ class AuthenticatorTest extends TestCase
     use TestDatabase;
     use RefreshDatabase;
     use withTestUser;
+    use withDatabaseSessionHandler;
 
     /**
      * Setup the test database.
@@ -78,6 +81,54 @@ class AuthenticatorTest extends TestCase
         $key = $this->ci->config['session.keys.current_user_id'];
         $this->assertNull($this->ci->session[$key]);
         $this->assertNotSame($testUser->id, $this->ci->session[$key]);
+    }
+
+    /**
+     * @depends testConstructor
+     * @param Authenticator $authenticator
+     */
+    public function testLoginWithSessionDatabase(Authenticator $authenticator)
+    {
+        // Reset CI Session
+        $this->useDatabaseSessionHandler();
+
+        // Create a test user
+        $testUser = $this->createTestUser();
+
+        // Check the table
+        $this->assertSame(0, SessionTable::count());
+
+        // Test session to avoid false positive
+        $key = $this->ci->config['session.keys.current_user_id'];
+        $this->assertNull($this->ci->session[$key]);
+        $this->assertNotSame($testUser->id, $this->ci->session[$key]);
+
+        // Login the test user
+        $authenticator->login($testUser, false);
+
+        // Test session to see if user was logged in
+        $this->assertNotNull($this->ci->session[$key]);
+        $this->assertSame($testUser->id, $this->ci->session[$key]);
+
+        // Close session to initiate write
+        session_write_close();
+
+        // Check the table again
+        $this->assertSame(1, SessionTable::count());
+
+        // Reopen session
+        $this->ci->session->start();
+
+        // Must logout to avoid test issue
+        $authenticator->logout(true);
+
+        // We'll test the logout system works too while we're at it (and depend on it)
+        $key = $this->ci->config['session.keys.current_user_id'];
+        $this->assertNull($this->ci->session[$key]);
+        $this->assertNotSame($testUser->id, $this->ci->session[$key]);
+
+        // Make sure table entry has been removed
+        $this->assertSame(0, SessionTable::count());
     }
 
     /**