+++ app/tests/TestCase.php
@@ -109,6 +109,9 @@ class TestCase extends BaseTestCase
                 call_user_func($callback);
             }
 
+            // Force destroy test sessions 
+            $this->ci->session->destroy();
+
             $this->ci = null;
         }