+++ CHANGELOG.md
@@ -7,12 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [v4.2.1]
 
+### Added
+- `UserFrosting\Sprinkle\Core\Database\Models\Session` model for the `sessions` db table.
+- `TEST_SESSION_HANDLER` environment variable to set the session save handler to use for Testing.
+- `withDatabaseSessionHandler` Trait for testing. Use `$this->useDatabaseSessionHandler()` to use database session handler in tests.
+
 ### Fixed
 - Italian translation ([#950])
 - User Registration failing when trying to register two accounts with the same email address ([#953])
 - Bad test case for `CoreController::getAsset`.
 - User Model `forceDelete` doesn't remove the record from the DB ([#951])
 - Fix PHP Fatal error that can be thrown when registering a new User
+- Session not working with database handler ([#952])
 
 ## [v4.2.0]
 ### Changed Requirements
@@ -715,6 +721,7 @@ See [http://learn.userfrosting.com/upgrading/40-to-41](Upgrading 4.0.x to 4.1.x 
 [#940]: https://github.com/userfrosting/UserFrosting/issues/940
 [#950]: https://github.com/userfrosting/UserFrosting/issues/950
 [#951]: https://github.com/userfrosting/UserFrosting/issues/951
+[#952]: https://github.com/userfrosting/UserFrosting/issues/952
 [#953]: https://github.com/userfrosting/UserFrosting/issues/953
 
 [v4.2.0]: https://github.com/userfrosting/UserFrosting/compare/v4.1.22...v4.2.0